server/pyproject.toml
[project]
name = "ragkit-server"
version = "0.1.0"
description = "ragkit FastAPI server: REST API, ingestion, embeddings, Qdrant, LCEL RAG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.35",
    "psycopg[binary]>=3.2.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "qdrant-client>=1.12.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "pymupdf>=1.24.0",
    "unstructured[docx,pptx,xlsx]>=0.15.0",
    "httpx>=0.27.0",
    "email-validator>=2.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]
# B008: FastAPI uses function calls (File/Form/Depends) in defaults by design.
# UP042: `(str, enum.Enum)` mixes well with pydantic; sticking with it.
ignore = ["E501", "B008", "UP042"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
